third_party/WebKit/LayoutTests/platform/linux/virtual/wheelscrolllatching/fast/events/wheel/wheel-scroll-latching-on-scrollbar-expected.txt

CONSOLE ERROR: line 2487: Uncaught Error: assert_equals: The child div should be fully scrolled. expected 115 but got 114
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught Error: assert_equals: The child div should be fully scrolled. expected 115 but got 114
PASS Scrolling on the child div scrollbar doesn't propagate to the parent div 
Harness: the test ran to completion.
